requirements.txt
tk
tweepy
python-dotenv
spellchecker
